--- src/com/getaji/bmshashwatcher/MainWindowModel.java
@@ -1,4 +1,0 @@
-package com.getaji.bmshashwatcher;
-
-public class MainWindowModel {
-}
