+++ src/com/getaji/bmshashwatcher/fxml/MainWindow.fxml
@@ -36,7 +36,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 <VBox prefHeight="488.0" prefWidth="779.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="com.getaji.bmshashwatcher.MainWindowController">
+      fx:controller="com.getaji.bmshashwatcher.controller.MainWindowController">
     <children>
         <MenuBar VBox.vgrow="NEVER">
             <menus>
