+++ src/com/getaji/bmshashwatcher/model/MainWindowModel.java
@@ -1,0 +1,4 @@
+package com.getaji.bmshashwatcher.model;
+
+public class MainWindowModel {
+}
