+++ src/com/getaji/bmshashwatcher/model/SongData.java
@@ -1,4 +1,4 @@
-package com.getaji.bmshashwatcher;
+package com.getaji.bmshashwatcher.model;
 
 public record SongData(String md5, String sha256, String title, String subtitle) {
