+++ src/com/getaji/bmshashwatcher/model/SupportedHashType.java
@@ -1,4 +1,4 @@
-package com.getaji.bmshashwatcher;
+package com.getaji.bmshashwatcher.model;
 
 public enum SupportedHashType {
     MD5,
